% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{build_coauthor_edges}
\alias{build_coauthor_edges}
\title{Build coauthorship edge list}
\usage{
build_coauthor_edges(works)
}
\arguments{
\item{works}{data.frame from OpenAlex}
}
\description{
Build coauthorship edge list
}
