% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{removes preprint authors from tibble of potential reviewers}
\usage{
clean(authors, candidates)
}
\arguments{
\item{authors}{the preprint authors}

\item{candidates}{the candidate preprint reviewers}
}
\description{
removes preprint authors from tibble of potential reviewers
}
