% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_intitutions.R
\name{get_author_institution}
\alias{get_author_institution}
\title{Fetch institutions for an author from OpenAlex}
\usage{
get_author_institution(author_id, verbose = TRUE)
}
\arguments{
\item{author_id}{OpenAlex author ID (e.g. "A123456789")}

\item{verbose}{logical}
}
\description{
Fetch institutions for an author from OpenAlex
}
