% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preprint.R
\name{get_preprint}
\alias{get_preprint}
\title{simple wrapper around openalexR to retrieve a preprint based on DOI.}
\usage{
get_preprint(doi)
}
\arguments{
\item{doi}{the preprint doi}
}
\description{
simple wrapper around openalexR to retrieve a preprint based on DOI.
}
