% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{parse_jane}
\alias{parse_jane}
\title{format JANE names to match OpenAlex}
\usage{
parse_jane(x)
}
\arguments{
\item{jane}{the name in the JANE db}

\item{alex}{the name in the OpenAlex db}
}
\description{
format JANE names to match OpenAlex
}
