% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarzan.R
\name{tarzan}
\alias{tarzan}
\title{tarzan queries JANE's API to retrieve the results of a JANE search. It then queries OpenAlex's API to retrieve the ORCID and affiliation for authors, based on the first PubMed paper detected by JANE.}
\usage{
tarzan(abstract, n)
}
\arguments{
\item{abstract}{the text to submit to JANE; for PCI, it is the abstract}

\item{n}{is the number of authors returned from JANE that you want to keep (default=10, max=100)}
}
\description{
tarzan queries JANE's API to retrieve the results of a JANE search. It then queries OpenAlex's API to retrieve the ORCID and affiliation for authors, based on the first PubMed paper detected by JANE.
}
