% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi.R
\name{detect_coi_pair}
\alias{detect_coi_pair}
\title{Detect conflict of interest between author and reviewer}
\usage{
detect_coi_pair(author_id, reviewer_id, verbose = FALSE)
}
\arguments{
\item{author_id}{OpenAlex author ID}

\item{reviewer_id}{OpenAlex author ID}

\item{verbose}{logical}
}
\description{
Detect conflict of interest between author and reviewer
}
