% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{get_author_works}
\alias{get_author_works}
\title{Fetch works for an author from OpenAlex}
\usage{
get_author_works(author_id, verbose = FALSE)
}
\arguments{
\item{author_id}{OpenAlex author ID (e.g. "A123456789")}

\item{verbose}{logical}
}
\description{
Fetch works for an author from OpenAlex
}
