% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_reviewers.R
\name{screen_reviewers}
\alias{screen_reviewers}
\title{Screen multiple reviewers for one author}
\usage{
screen_reviewers(author_id, reviewer_ids)
}
\description{
Screen multiple reviewers for one author
}
