% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarzan.R
\name{tarzan}
\alias{tarzan}
\title{tarzan queries JANE's APÏ}
\usage{
tarzan(abstract)
}
\arguments{
\item{abstract}{the text to submit to JANE; for PCI, it is the abstract}
}
\description{
tarzan queries JANE's APÏ
}
